requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
sentence-transformers
huggingface_hub
langchain-community
einops
playwright
